requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.2
pytest>=7.4.0

# Progress bars

tqdm>=4.65.0

# NLP / RAG
transformers>=4.44.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
nltk>=3.8.1
huggingface-hub>=0.24.0
datasets>=2.20.0

# Torch (CPU)
torch>=2.2.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Utilities
pyyaml>=6.0.1
ujson>=5.10.0
